+++ climaf/derived_variables/ipsl_derived_variables.py
@@ -4,7 +4,7 @@
 from __future__ import print_function, division, unicode_literals, absolute_import
 
 from env.environment import *
-from climaf.api import calias
+from climaf.driver import calias
 from climaf.operators_derive import derive
 
 # Content was sent to various places : prokect IGCM_OUT, atmposphere_derived_variables, ocean_derived_variables
