+++ climaf/projects/cmip6_valid.py
@@ -1,3 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from __future__ import print_function, division, unicode_literals, absolute_import
+
+
 from env.environment import cprojects
