+++ climaf/projects/cruts3.py
@@ -22,9 +22,10 @@ Example of an 'cruts3' project dataset declaration ::
 from __future__ import print_function, division, unicode_literals, absolute_import
 
 from env.environment import *
+from env.site_settings import atCNRM
+
 from climaf.dataloc import dataloc
 from climaf.classes import cproject, calias, cdef
-from env.site_settings import atCNRM
 
 if atCNRM:
     cproject('cruts3', 'grid')  # no grid writes as '', otherwise as e.g. 'T127'
